refactor: extract semver utils to separate file
