+++ lib/dependency-versions.ts
@@ -1,6 +1,13 @@
-import semver from 'semver';
 import editJsonFile from 'edit-json-file';
 import { Package } from './package.js';
+import {
+  compareVersionRanges,
+  compareVersionRangesSafe,
+  versionRangeToRange,
+  getLatestVersion,
+  getHighestRangeType,
+} from './semver.js';
+import semver from 'semver';
 
 export type DependenciesToVersionsSeen = Map<
   string,
@@ -340,67 +347,4 @@ export function fixMismatchingVersions(
     fixed,
     notFixed,
   };
-}
-
-// This version doesn't throw for when we want to ignore invalid versions that might be present.
-export function compareVersionRangesSafe(a: string, b: string): 0 | -1 | 1 {
-  try {
-    return compareVersionRanges(a, b);
-  } catch {
-    return 0;
-  }
-}
-
-// Compare semver version ranges like ^1.0.0, ~2.5.0, 3.0.0, etc.
-export function compareVersionRanges(a: string, b: string): 0 | -1 | 1 {
-  // Coerce to normalized version without any range prefix.
-  const aVersion = semver.coerce(a);
-  const bVersion = semver.coerce(b);
-  if (!aVersion) {
-    throw new Error(`Invalid Version: ${a}`);
-  }
-  if (!bVersion) {
-    throw new Error(`Invalid Version: ${b}`);
-  }
-
-  if (semver.eq(aVersion, bVersion)) {
-    // Same version, so decide which range is considered higher.
-    const aRange = versionRangeToRange(a);
-    const bRange = versionRangeToRange(b);
-    return compareRanges(aRange, bRange);
-  }
-
-  // Greater version considered higher.
-  return semver.gt(aVersion, bVersion) ? 1 : -1;
-}
-
-const RANGE_PRECEDENCE = ['~', '^']; // Lowest to highest.
-
-// Compare semver ranges like ^, ~, etc.
-export function compareRanges(aRange: string, bRange: string): 0 | -1 | 1 {
-  const aRangePrecedence = RANGE_PRECEDENCE.indexOf(aRange);
-  const bRangePrecedence = RANGE_PRECEDENCE.indexOf(bRange);
-  if (aRangePrecedence > bRangePrecedence) {
-    return 1;
-  } else if (aRangePrecedence < bRangePrecedence) {
-    return -1;
-  }
-  return 0;
-}
-
-// Example input: ^1.0.0, output: ^
-export function versionRangeToRange(versionRange: string): string {
-  const match = versionRange.match(/^\D+/);
-  return match ? match[0] : '';
-}
-
-export function getLatestVersion(versions: string[]): string {
-  const sortedVersions = versions.sort(compareVersionRanges);
-  return sortedVersions[sortedVersions.length - 1]; // Latest version will be sorted to end of list.
-}
-
-// Example input: ['~', '^'], output: '^'
-export function getHighestRangeType(ranges: string[]): string {
-  const sorted = ranges.sort(compareRanges);
-  return sorted[sorted.length - 1]; // Range with highest precedence will be sorted to end of list.
 }
